Have removing recommendation working
