+++ src/main/javascript/book-master/src/pages/RecommendationSuccess.js
@@ -5,7 +5,7 @@ const RecommendationSuccess = () =>{
     return(
         <>
             <div className='container mt-3'>
-                <h1>Recommendation Added</h1>
+                <h1>Recommendation Transaction Completed</h1>
                 <a href="/"><h3>Return Home</h3></a>
                 <a href="/library_search"><h3>Look At Other Books</h3></a>
             </div>
